Update: move `RegisterDomainCoordinator` to its own file
